Moved java shim behave test to the existing file
