+++ bddtests/steps/peer_basic_impl.py
@@ -152,6 +152,42 @@ def step_impl(context, seconds):
 def step_impl(context, seconds):
     time.sleep(float(seconds))
 
+@when(u'I deploy lang chaincode "{chaincodePath}" of "{chainLang}" with ctor "{ctor}" to "{containerName}"')
+def step_impl(context, chaincodePath, chainLang, ctor, containerName):
+    print("Printing chaincode language " + chainLang)
+    ipAddress = bdd_test_util.ipFromContainerNamePart(containerName, context.compose_containers)
+    request_url = buildUrl(context, ipAddress, "/devops/deploy")
+    print("Requesting path = {0}".format(request_url))
+    args = []
+    if 'table' in context:
+       # There is ctor arguments
+       args = context.table[0].cells
+    #typeGolang =
+
+    # Create a ChaincodeSpec structure
+    chaincodeSpec = {
+        "type": chainLang,
+        "chaincodeID": {
+            "path" : chaincodePath,
+            "name" : ""
+        },
+        "ctorMsg":  {
+            "function" : ctor,
+            "args" : args
+        },
+    }
+    if 'userName' in context:
+        chaincodeSpec["secureContext"] = context.userName
+
+    resp = requests.post(request_url, headers={'Content-type': 'application/json'}, data=json.dumps(chaincodeSpec), verify=False)
+    assert resp.status_code == 200, "Failed to POST to %s:  %s" %(request_url, resp.text)
+    context.response = resp
+    chaincodeName = resp.json()['message']
+    chaincodeSpec['chaincodeID']['name'] = chaincodeName
+    context.chaincodeSpec = chaincodeSpec
+    print(json.dumps(chaincodeSpec, indent=4))
+    print("")
+
 
 @when(u'I deploy chaincode "{chaincodePath}" with ctor "{ctor}" to "{containerName}"')
 def step_impl(context, chaincodePath, ctor, containerName):
